# 949 Correct handling of dt_min in integrators