--- a/cpp/memilio/math/adapt_rk.cpp
+++ b/cpp/memilio/math/adapt_rk.cpp
@@ -87,6 +87,10 @@ bool RKIntegratorCore::step(const DerivFunction& f, Eigen::Ref<const Eigen::Vect
     m_yt_eval = yt;
 
     while (!converged && !failed_step_size_adapt) {
+        if (dt < m_dt_min) {
+            failed_step_size_adapt = true;
+            dt                     = m_dt_min;
+        }
         // compute first column of kt, i.e. kt_0 for each y in yt_eval
         f(m_yt_eval, t, m_kt_values.col(0));
 
@@ -113,7 +117,7 @@ bool RKIntegratorCore::step(const DerivFunction& f, Eigen::Ref<const Eigen::Vect
 
         converged = (m_error_estimate <= m_eps).all(); // convergence criterion
 
-        if (converged) {
+        if (converged || failed_step_size_adapt) {
             // if sufficiently exact, return ytp1, which currently contains the lower order approximation
             // (higher order is not always higher accuracy)
             t += dt; // this is the t where ytp1 belongs to
@@ -128,13 +132,9 @@ bool RKIntegratorCore::step(const DerivFunction& f, Eigen::Ref<const Eigen::Vect
         // and to avoid dt_new -> dt for step decreases when |error_estimate - eps| -> 0
         dt_new *= 0.9;
         // check if updated dt stays within desired bounds and update dt for next step
-        if (m_dt_min < dt_new) {
-            dt = std::min(dt_new, m_dt_max);
-        }
-        else {
-            failed_step_size_adapt = true;
-        }
+        dt = std::min(dt_new, m_dt_max);
     }
+    dt = std::max(dt, m_dt_min);
     return !failed_step_size_adapt;
 }