--- a/cpp/memilio/math/stepper_wrapper.h
+++ b/cpp/memilio/math/stepper_wrapper.h
@@ -32,7 +32,7 @@ MSVC_WARNING_DISABLE_PUSH(4127)
 #include "boost/numeric/odeint/stepper/runge_kutta4.hpp"
 #include "boost/numeric/odeint/stepper/runge_kutta_fehlberg78.hpp"
 #include "boost/numeric/odeint/stepper/runge_kutta_cash_karp54.hpp"
-#include "boost/numeric/odeint/stepper/runge_kutta_dopri5.hpp"
+// #include "boost/numeric/odeint/stepper/runge_kutta_dopri5.hpp" // TODO: reenable once boost bug is fixed
 MSVC_WARNING_POP()
 GCC_CLANG_DIAGNOSTIC(pop)
 
@@ -50,6 +50,12 @@ template <template <class State = Eigen::VectorXd, class Value = double, class D
           class ControlledStepper>
 class ControlledStepperWrapper : public mio::IntegratorCore
 {
+    using Stepper                         = boost::numeric::odeint::controlled_runge_kutta<ControlledStepper<>>;
+    static constexpr bool is_fsal_stepper = std::is_same_v<typename Stepper::stepper_type::stepper_category,
+                                                           boost::numeric::odeint::explicit_error_stepper_fsal_tag>;
+    static_assert(!is_fsal_stepper,
+                  "FSAL steppers cannot be used unitl https://github.com/boostorg/odeint/issues/72 is resolved.");
+
 public:
     /**
      * @brief Set up the integrator
@@ -70,32 +76,59 @@ public:
     }
 
     /**
-    * @brief Make a single integration step of a system of ODEs and adapt step width
-    * @param[in] yt value of y at t, y(t)
-    * @param[in,out] t current time step h=dt
-    * @param[in,out] dt current time step h=dt
-    * @param[out] ytp1 approximated value y(t+1)
+    * @brief Make a single integration step of a system of ODEs and adapt the step size dt.
+    * @param[in] yt Value of y at t_{k}, y(t_{k}).
+    * @param[in,out] t Current time step t_{k} for some k. Will be set to t_{k+1} in [t_{k} + dt_min, t_{k} + dt].
+    * @param[in,out] dt Current time step size h=dt. Overwritten by an estimated optimal step size for the next step.
+    * @param[out] ytp1 The approximated value of y(t_{k+1}).
     */
     bool step(const mio::DerivFunction& f, Eigen::Ref<Eigen::VectorXd const> yt, double& t, double& dt,
               Eigen::Ref<Eigen::VectorXd> ytp1) const override
     {
-        // copy y(t) to dydt, to retrieve the VectorXd from the Ref
-        dydt               = yt;
-        const double t_old = t; // t is updated by try_step on a successfull step
-        do {
-            // we use the scheme try_step(sys, inout, t, dt) with sys=f, inout=y(t) for
-            // in-place computation. This is similiar to do_step, but it can update t and dt
-            m_stepper.try_step(
-                // reorder arguments of the DerivFunction f for the stepper
-                [&](const Eigen::VectorXd& x, Eigen::VectorXd& dxds, double s) {
-                    dxds.resizeLike(x); // try_step calls sys with a vector of size 0 for some reason
-                    f(x, s, dxds);
-                },
-                dydt, t, dt);
-            // stop on a successfull step or a failed step size adaption (w.r.t. the minimal step size)
-        } while (t == t_old && dt > m_dt_min);
-        ytp1 = dydt; // output new y(t)
-        return dt > m_dt_min;
+        using boost::numeric::odeint::fail;
+        assert(0 <= m_dt_min);
+        assert(m_dt_min <= m_dt_max);
+        // set initial values for exit conditions
+        auto step_result = fail;
+        bool is_dt_valid = true;
+        // copy vectors from the references, since the stepper cannot (trivially) handle Eigen::Ref
+        m_ytp1 = ytp1;
+        m_yt   = yt;
+        // make a integration step, adapting dt to a possibly larger value on success,
+        // or a strictly smaller value on fail.
+        // stop only on a successful step or a failed step size adaption (w.r.t. the minimal step size m_dt_min)
+        while (step_result == fail && is_dt_valid) {
+            if (dt < m_dt_min) {
+                is_dt_valid = false;
+                dt          = m_dt_min;
+            }
+            // we use the scheme try_step(sys, in, t, out, dt) with sys=f, in=y(t_{k}), out=y(t_{k+1}).
+            // this is similiar to do_step, but it can adapt the step size dt. if successfull, it also updates t.
+            if constexpr (!is_fsal_stepper) { // prevent compile time errors with fsal steppers
+                step_result = m_stepper.try_step(
+                    // reorder arguments of the DerivFunction f for the stepper
+                    [&](const Eigen::VectorXd& x, Eigen::VectorXd& dxds, double s) {
+                        dxds.resizeLike(x); // boost resizers cannot resize Eigen::Vector, hence we need to do that here
+                        f(x, s, dxds);
+                    },
+                    m_yt, t, m_ytp1, dt);
+            }
+        }
+        // output the new value by copying it back to the reference
+        ytp1 = m_ytp1;
+        // bound dt from below
+        // the last adaptive step (successfull or not) may have calculated a new step size smaller than m_dt_min
+        dt = std::max(dt, m_dt_min);
+        // check whether the last step failed (which means that m_dt_min was still too large to suffice tolerances)
+        if (step_result == fail) {
+            // adaptive stepping failed, but we still return the result of the last attempt
+            t += m_dt_min;
+            return false;
+        }
+        else {
+            // successfully made an integration step
+            return true;
+        }
     }
 
     /// @param tol the required absolute tolerance for comparison of the iterative approximation
@@ -126,21 +159,18 @@ public:
     }
 
 private:
-    boost::numeric::odeint::controlled_runge_kutta<ControlledStepper<>> create_stepper()
+    /// @brief (Re)initialize the internal stepper.
+    Stepper create_stepper()
     {
         // for more options see: boost/boost/numeric/odeint/stepper/controlled_runge_kutta.hpp
-        return boost::numeric::odeint::controlled_runge_kutta<ControlledStepper<>>(
-            boost::numeric::odeint::default_error_checker<typename ControlledStepper<>::value_type,
-                                                          typename ControlledStepper<>::algebra_type,
-                                                          typename ControlledStepper<>::operations_type>(m_abs_tol,
-                                                                                                         m_rel_tol),
-            boost::numeric::odeint::default_step_adjuster<typename ControlledStepper<>::value_type,
-                                                          typename ControlledStepper<>::time_type>(m_dt_max));
+        return Stepper(typename Stepper::error_checker_type(m_abs_tol, m_rel_tol),
+                       typename Stepper::step_adjuster_type(m_dt_max));
     }
 
-    double m_abs_tol, m_rel_tol, m_dt_min, m_dt_max; // integrator parameters
-    mutable Eigen::VectorXd dydt;
-    mutable boost::numeric::odeint::controlled_runge_kutta<ControlledStepper<>> m_stepper;
+    double m_abs_tol, m_rel_tol; ///< Absolute and relative tolerances for integration.
+    double m_dt_min, m_dt_max; ///< Lower and upper bound to the step size dt.
+    mutable Eigen::VectorXd m_ytp1, m_yt; ///< Temporary storage to avoid allocations in step function.
+    mutable Stepper m_stepper; ///< A stepper instance used for integration.
 };
 
 } // namespace mio