--- a/cpp/tests/test_odesecir.cpp
+++ b/cpp/tests/test_odesecir.cpp
@@ -87,7 +87,7 @@ TEST(TestOdeSecir, compareWithPreviousRun)
     auto integrator = std::make_shared<mio::RKIntegratorCore>();
     integrator->set_dt_min(0.3);
     integrator->set_dt_max(1.0);
-    integrator->set_rel_tolerance(1e-4);
+    integrator->set_rel_tolerance(1e-3);
     integrator->set_abs_tolerance(1e-1);
     mio::TimeSeries<double> secihurd = simulate(t0, tmax, dt, model, integrator);