@reneSchm 2024-03-08
So I added a small warning when the given dt was out of bounds, and strangely enough it might have improved performance for some steppers?
- `RKIntegratorCore` aka adapt_rk and  `ControlledStepperWrapper<runge_kutta_fehlberg78>` aka rkf78 are consistently about 50ns faster. Which is nice, but confusing. Apart from the couple lines for the logging, the assembly looked like only some addresses got reordered, at least I didn't see any significant changes. My best guess is that I got lucky and some reordering caused better cache hits, but I don't know.
- `ControlledStepperWrapper<runge_kutta_cash_carp54>` aka rk_ck54 seems to be roughly 10ns slower, but that is within the margin of error for these measurements. This is as expected, and I think well acceptable.

Before logging (dae153c):
```
simulate SecirModel adapt_rk            1008 ns         1007 ns       696182
simulate SecirModel boost rk_ck54        746 ns          746 ns       938309
simulate SecirModel boost rkf78         2196 ns         2195 ns       317635
```
With logging (b93c0e6):
```
simulate SecirModel adapt_rk             935 ns          934 ns       749251
simulate SecirModel boost rk_ck54        759 ns          759 ns       921252
simulate SecirModel boost rkf78         2141 ns         2141 ns       323717
```